fix if condition in `get_column_count`
`get_column_count` checked for `column_count > 1` instead of
`column_count > 0`.  It returned `None` for single-column
specifications.  This was a bug, although it did not lead to wrong
behavior because in this situation, `make_template` imputed
`len(json_filenames)` for the column count.  Assuming that the number of
JSON files passed to `tomltable` on the command line is the same as the
number of columns in the specification, the two ways of calculating the
column count are consistent with each other.